third_party/blink/web_tests/svg/dynamic-updates/SVGImageElement-dom-height-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'height' attribute of the SVGImageElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS imageElement.getAttribute('height') is "100"
PASS imageElement.getAttribute('height') is "200"
PASS successfullyParsed is true

TEST COMPLETE
